chore: more re-organizing
